Project Commit
This Asp.net Core Console application creates an HTTP client. It asks the user to add a url, then sends a HTTP Request to the url. It checks if the url starts with https:// or http:// and makes a get request. It reads the html response from the server and saves it in a .txt file inside the project.